Set timeout to repeat input on key-press to 250ms
I struggled to distinguish between a simple key-stroke and a pressed
key, which should lead to repeat inputs. Initially I copied this value
from my macos-settings where I use the same keyboard, but it turns out
that the timing is somehow different.

Remember set this to 200ms on macos and 250ms on X11.

Note, I have no idea which one is more accurate, but since everything on
a mac feels (and is) slow I would guess linux is correct and apple lies.