Initialize zoxide correctly, so I get proper completions
I used zoxide wrong all the time... By simply initiating it correctly it
immediately get's smarter and better. Imagine you have a dir like this:

  ~/deep/level/nested/but-with-very-specific-name/xy11-with-many-more

Imagine further that in your current directory is nothing that starts
with `xy`, then you only have to enter `cd xy` and zoxide will figure
out that you meant the deeply nested complicated name and change the
directory to there! - How awesome is that!

So, keep in mind use zoxide correct and it becomes easy to use!