Configure url-views in neomutt mails via urlscan and urlview
Some mails are just stupidly hard to parse, especially if the HTML
contains dozens of images. Kununu even managed to include the same
amount of images into the text-mails with all their marketing foo in the
URL. So, since these were so hard to understand I needed multiple
different "grades" of URL lists. This change allow be to choose between
different views.